build: fix warning re comparison of enum and uint
The warning is observed when tarantool is compiled by GCC 9.1.0.

Warnings are treated as errors during a debug build or when
-DENABLE_WERROR=ON option is passed to cmake, that is usual for our
testing jobs in CI.

The commit that introduces the problem is
3a8adccf5fee5b8a19583c8eb846b18f10e449d6 ('access: fix invalid error
type for not found user').

+++ src/box/user.cc
@@ -530,7 +530,7 @@ user_find_by_name(const char *name, uint32_t len)
 			return user;
 	}
 	diag_set(ClientError, ER_NO_SUCH_USER,
-		 tt_cstr(name, MIN(BOX_INVALID_NAME_MAX, len)));
+		 tt_cstr(name, MIN((uint32_t) BOX_INVALID_NAME_MAX, len)));
 	return NULL;
 }
